webview

+++ app/src/main/java/com/scruge/scruge/model/entity/Economics.kt
@@ -1,6 +1,6 @@
 package com.scruge.scruge.model.entity
 
-data class Economics(val hardCap:Double, val softCap:Double, val raised:Double,
+data class Economics(val hardCap:Int, val softCap:Int, val raised:Double,
                      val publicTokenPercent:Double, val tokenSupply:Double,
                      val annualInflationPercent:Range,
                      val minUserContribution:Double, val maxUserContribution:Double,
